babix2/requirements.txt
certifi==2017.11.5
chardet==3.0.4
dill==0.2.5
future==0.16.0
idna==2.6
numpy==1.13.3
PyExecJS==1.5.0
PyFunctional==1.0.0
requests==2.18.4
six==1.10.0
tabulate==0.7.7
urllib3==1.22
